requirements.txt
pandas>=1.5.0
openai>=1.0.0
tqdm>=4.64.0
tenacity>=8.0.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0
python-dotenv>=1.0.0
